\name{queryResult}
\alias{queryResult}
\docType{data}
\title{
Travistorrent
}
\description{
Dataset
}
\usage{data("queryResult")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{author_mail}}{a numeric vector}
    \item{\code{gh_project_name}}{a numeric vector}
    \item{\code{git_commit}}{a numeric vector}
    \item{\code{tr_build_id}}{a numeric vector}
    \item{\code{tr_status}}{a numeric vector}
  }
}
\examples{
## data(queryResult)
## maybe str(queryResult) ; plot(queryResult) ...
}
\keyword{datasets}
